Support measuring build operations in the profiler report
