+++ subprojects/internal-performance-testing/src/main/groovy/org/gradle/performance/fixture/GradleBuildExperimentSpec.groovy
@@ -26,11 +26,13 @@ import java.util.function.Function
 class GradleBuildExperimentSpec extends BuildExperimentSpec {
     final GradleInvocationSpec invocation
     final ImmutableList<Function<InvocationSettings, BuildMutator>> buildMutators
+    final ImmutableList<String> measuredBuildOperations
 
-    GradleBuildExperimentSpec(String displayName, String projectName, File workingDirectory, GradleInvocationSpec invocation, Integer warmUpCount, Integer invocationCount, BuildExperimentListener listener, InvocationCustomizer invocationCustomizer, List<Function<InvocationSettings, BuildMutator>> buildMutators) {
+    GradleBuildExperimentSpec(String displayName, String projectName, File workingDirectory, GradleInvocationSpec invocation, Integer warmUpCount, Integer invocationCount, BuildExperimentListener listener, InvocationCustomizer invocationCustomizer, List<Function<InvocationSettings, BuildMutator>> buildMutators, ImmutableList<String> measuredBuildOperations) {
         super(displayName, projectName, workingDirectory, warmUpCount, invocationCount, listener, invocationCustomizer)
         this.invocation = invocation
         this.buildMutators = buildMutators
+        this.measuredBuildOperations = measuredBuildOperations
     }
 
     static GradleBuilder builder() {
@@ -51,6 +53,7 @@ class GradleBuildExperimentSpec extends BuildExperimentSpec {
         Integer invocationCount
         BuildExperimentListener listener
         List<Function<InvocationSettings, BuildMutator>> buildMutators
+        ImmutableList<String> measuredBuildOperations
         InvocationCustomizer invocationCustomizer
 
         GradleBuilder displayName(String displayName) {
@@ -88,6 +91,11 @@ class GradleBuildExperimentSpec extends BuildExperimentSpec {
             this
         }
 
+        GradleBuilder measuredBuildOperations(List<String> measuredBuildOperations) {
+            this.measuredBuildOperations = ImmutableList.copyOf(measuredBuildOperations)
+            this
+        }
+
         GradleBuilder invocationCustomizer(InvocationCustomizer invocationCustomizer) {
             this.invocationCustomizer = invocationCustomizer
             this
@@ -98,7 +106,7 @@ class GradleBuildExperimentSpec extends BuildExperimentSpec {
             assert displayName != null
             assert invocation != null
 
-            new GradleBuildExperimentSpec(displayName, projectName, workingDirectory, invocation.build(), warmUpCount, invocationCount, listener, invocationCustomizer, buildMutators)
+            new GradleBuildExperimentSpec(displayName, projectName, workingDirectory, invocation.build(), warmUpCount, invocationCount, listener, invocationCustomizer, buildMutators, measuredBuildOperations)
         }
     }
 }
